# backend/documentation/swagger.yaml
...
swagger: "2.0"
info:
  title: "TRENDMART API"
  description: "API documentation for the TRENDMART application."
  version: "1.0.0"
host: "localhost:8000"
basePath: "/"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# =========================
# Paths
# =========================
paths:
...
  /payments/config:
    get:
      tags:
        - Payments
      summary: Get Stripe publishable key
      responses:
        200:
          description: Publishable key
          schema:
            type: object
            properties:
              publishableKey:
                type: string
        500:
          description: Config error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /payments/intent:
    post:
      tags:
        - Payments
      summary: Create or reuse a Stripe PaymentIntent for an order
      description: Creates a PaymentIntent for the given `order_id`. If a non-final intent exists, returns its `client_secret`.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/PaymentIntentCreate'
      responses:
        200:
          description: Intent created or reused
          schema:
            $ref: '#/definitions/PaymentIntentResponse'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        502:
          description: Stripe error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /payments/webhook:
    post:
      tags:
        - Payments
      summary: Stripe webhook endpoint
      responses:
        200:
          description: Processed
          schema:
            type: object
            properties:
              status:
                type: string
                example: ok
        400:
          description: Invalid payload or signature
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Config error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /payments/{payment_id}:
    get:
      tags:
        - Payments
      summary: Get payment status by internal id
      parameters:
        - in: path
          name: payment_id
          required: true
          type: integer
      responses:
        200:
          description: Payment status
          schema:
            $ref: "#/definitions/PaymentStatusResponse"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /payments/by-order/{order_id}:
    get:
      tags:
        - Payments
      summary: Get payment status by order id
      parameters:
        - in: path
          name: order_id
          required: true
          type: integer
      responses:
        200:
          description: Payment status
          schema:
            $ref: "#/definitions/PaymentStatusResponse"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /payments/{payment_id}/refund:
    post:
      tags:
        - Payments
      summary: Create a refund
      parameters:
        - in: path
          name: payment_id
          required: true
          type: integer
        - in: body
          name: body
          required: false
          schema:
            $ref: "#/definitions/RefundCreate"
      responses:
        201:
          description: Refund created
          schema:
            $ref: "#/definitions/Refund"
        400:
          description: Stripe error or bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Payment not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /payments/{payment_id}/refunds:
    get:
      tags:
        - Payments
      summary: List refunds for a payment
      parameters:
        - in: path
          name: payment_id
          required: true
          type: integer
      responses:
        200:
          description: Refunds list
          schema:
            $ref: "#/definitions/RefundListResponse"
        404:
          description: Payment not found
          schema:
            $ref: "#/definitions/ErrorResponse"
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter 'Bearer' followed by your JWT token. Example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."

# =========================
# Definitions (Schemas)
# =========================
definitions:
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: "Error message"
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Detailed error message"
      details:
        type: object
        description: "Validation error details"

  # -----------------
  # Payments Schemas
  # -----------------
  PaymentIntentCreate:
    type: object
    required: [order_id]
    properties:
      order_id:
        type: integer
      currency:
        type: string
        example: usd
      metadata:
        type: object
        description: Optional metadata; user_id automatically included when available.
  PaymentIntentResponse:
    type: object
    properties:
      client_secret:
        type: string
      payment_intent_id:
        type: string
      payment_id:
        type: integer
      order_id:
        type: integer
      status:
        type: string
        example: pending
      amount_cents:
        type: integer
        description: Convenience amount in cents (derived)
      total_amount:
        type: number
        format: float
        description: Payment amount in dollars
  PaymentStatusResponse:
    type: object
    properties:
      payment_id:
        type: integer
      order_id:
        type: integer
      stripe_payment_intent_id:
        type: string
      status:
        type: string
      amount_cents:
        type: integer
      total_amount:
        type: number
        format: float
  RefundCreate:
    type: object
    properties:
      amount_cents:
        type: integer
        description: Optional amount in cents for a partial refund
      reason:
        type: string
        description: Optional refund reason (e.g., requested_by_customer)
  Refund:
    type: object
    properties:
      id:
        type: string
      status:
        type: string
      amount:
        type: integer
      currency:
        type: string
      created:
        type: integer
      payment_intent_id:
        type: string
  RefundListResponse:
    type: object
    properties:
      refunds:
        type: array
        items:
          $ref: "#/definitions/Refund"

  # -----------------
  # Bulk User Data Schemas
  # -----------------
...